Move resizablecomponent to src root
