+++ lib/src/main/java/com/zetcode/resizablecomponent/Resizable.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.resizablecomponent;
 
 
 import javax.swing.JComponent;
